+++ composeApp/src/commonMain/kotlin/echo/music/enhanced/di/ViewModelModule.kt
@@ -87,6 +87,8 @@ val viewModelModule =
             HomeViewModel(
                 get(),
                 get(),
+                get(),
+                get(),
             )
         }
         viewModel {